lighter_client/requirements.txt
# Lighter交易所客户端依赖

# 核心依赖
lighter-python>=1.0.0  # Lighter官方Python SDK

# WebSocket支持
websockets>=12.0  # WebSocket客户端库

# 环境变量管理
python-dotenv>=1.0.0  # .env文件支持

# 异步支持
asyncio>=3.4.3  # Python内置异步库

# 日志和工具
logging>=0.4.9.6  # Python内置日志库

# 类型提示支持
typing-extensions>=4.0.0  # 类型提示扩展

# 开发依赖（可选）
pytest>=7.0.0  # 测试框架
pytest-asyncio>=0.21.0  # 异步测试支持
black>=23.0.0  # 代码格式化
flake8>=6.0.0  # 代码检查
mypy>=1.0.0  # 类型检查
